successful build almalinux-10 utm provider

+++ packer_templates/scripts/_common/build_tools.sh
@@ -47,12 +47,22 @@ virtualbox-iso|virtualbox-ovf)
     echo "Skipping Virtualbox guest additions installation on aarch64 architecture for opensuse and derivatives"
   fi
   ;;
+hyperv-iso)
+  echo "nothing to do for hyperv-iso"
+  ;;
+utm-iso|qemu)
+  echo "nothing to do for utm or qemu"
+  ;;
 *)
   echo "Unknown Packer Builder Type >>$PACKER_BUILDER_TYPE<< selected."
   exit 0
   ;;
 esac
 
-echo "build tools installed rebooting"
-shutdown -r now
-sleep 60
+if [ -f /var/run/reboot-required ] || ! command -v needs-restarting -r &> /dev/null || ! command -v needs-restarting -s &> /dev/null; then
+  echo "pkgs installed needing reboot"
+  shutdown -r now
+  sleep 60
+else
+  echo "no pkgs installed needing reboot"
+fi
