+++ packer_templates/scripts/_common/guest_tools_parallels.sh
@@ -44,7 +44,13 @@ parallels-iso|parallels-pvm|parallels-ipsw)
   else
     echo "Skipping Parallels Tools installation on aarch64 architecture for opensuse and derivatives"
   fi
-  shutdown -r now
-  sleep 60
+
+  if [ -f /var/run/reboot-required ] || ! command -v needs-restarting -r &> /dev/null || ! command -v needs-restarting -s &> /dev/null; then
+    echo "pkgs installed needing reboot"
+    shutdown -r now
+    sleep 60
+  else
+    echo "no pkgs installed needing reboot"
+  fi
   ;;
 esac
