+++ packer_templates/scripts/_common/update_packages.sh
@@ -86,6 +86,10 @@ else
   exit 1
 fi
 
-echo "updates installed rebooting"
-shutdown -r now
-sleep 60
+if [ -f /var/run/reboot-required ] || ! command -v needs-restarting -r &> /dev/null || ! command -v needs-restarting -s &> /dev/null; then
+  echo "pkgs installed needing reboot"
+  shutdown -r now
+  sleep 60
+else
+  echo "no pkgs installed needing reboot"
+fi
